// packages/next/src/server/dev/hot-reloader-types.ts
export const enum HMR_MESSAGE_SENT_TO_SERVER {
  // JSON messages:
  MCP_ERROR_STATE_RESPONSE = 'mcp-error-state-response',
  RUNTIME_ERROR_STATE = 'runtime-error-state',
  MCP_PAGE_METADATA_RESPONSE = 'mcp-page-metadata-response',
  PING = 'ping',
}
// ...
export interface RuntimeErrorStateError {
  id: number
  error: {
    name?: string
    message?: string
    stack?: string
  } | null
  frames: readonly {
    file: string | null
    methodName: string
    line1: number | null
    column1: number | null
  }[]
  type: 'runtime' | 'recoverable' | 'console'
  isFatal: boolean
}

export interface RuntimeErrorStateUpdate {
  event: HMR_MESSAGE_SENT_TO_SERVER.RUNTIME_ERROR_STATE
  clientId: number
  pathname: string
  errorState: {
    errors: readonly RuntimeErrorStateError[]
    routerType: 'app' | 'pages'
  }
}